requirements.txt
anthropic==0.39.0
langchain==0.3.0
langchain-anthropic==0.3.0
langgraph==0.2.45
pandas==2.2.0
numpy==1.26.0
matplotlib==3.8.0
seaborn==0.13.0
python-dotenv==1.0.0
httpx==0.27.0
requests==2.31.0